==> notebooks_stories/2_analyze/06_gemv_basic.ipynb <==
%load_ext autoreload
%autoreload 2
import flatmaps_helper


# main load
normalize_flatmaps = False
gemv_flatmaps_dict_S02, gemv_flatmaps_dict_S03 = flatmaps_helper.load_flatmaps(
    normalize_flatmaps, )

gemv_flatmaps_dict_S03
